requirements.txt
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==38.0.1
idna==3.4
pycparser==2.21
pyOpenSSL==22.0.0
requests==2.26.0
urllib3==1.26.8
dnspython==2.2.1